pyproject.toml
[tool.poetry]
name = "asobi"
version = "0.1.0"
description = ""
authors = ["maetin <maetin0324@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.27.7"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
